#144 [feat] : 삭제, 편집 팝업 구현
